ast: moved collect_vars visitor from address_expr_collector to its own module.
